remove root agent abstraction
