+++ packages/ai/src/agents/classes/root.ts
@@ -1,21 +1,60 @@
 import { ChatType, LLMProvider, OPENROUTER_MODELS, type ModelConfig } from '@onlook/models';
+import { stepCountIs, streamText, type ModelMessage, type ToolSet } from 'ai';
 import { getAskModeSystemPrompt, getCreatePageSystemPrompt, getSystemPrompt, getToolSetFromType, initModel } from '../../index';
-import { BaseAgent } from '../models/base';
 
-export class RootAgent extends BaseAgent {
+export class RootAgent {
     readonly id = 'root-agent';
     private readonly chatType: ChatType;
     readonly modelConfig: ModelConfig;
+    protected readonly toolSet: ToolSet;
 
     constructor(chatType: ChatType, modelConfig: ModelConfig) {
-        super(getToolSetFromType(chatType));
-
         this.chatType = chatType;
         this.modelConfig = modelConfig;
+        this.toolSet = getToolSetFromType(chatType);
     }
 
     get systemPrompt(): string {
         return this.getSystemPromptFromType(this.chatType);
+    }
+
+    getSystemPrompt() {
+        return this.systemPrompt;
+    }
+
+    getToolSet() {
+        return this.toolSet;
+    }
+
+    // Default streamText configuration - can be overridden by subclasses
+    protected getStreamTextConfig() {
+        return {
+            maxSteps: 20,
+        };
+    }
+
+    // Main streamText method that uses the agent's configuration
+    streamText(
+        messages: ModelMessage[] = [],
+        additionalConfig: Omit<Partial<Parameters<typeof streamText>[0]>, 'model' | 'headers' | 'tools' | 'stopWhen' | 'messages' | 'prompt'> = {},
+    ) {
+        const config = this.getStreamTextConfig();
+
+        return streamText({
+            model: this.modelConfig.model,
+            headers: this.modelConfig.headers,
+            tools: this.getToolSet(),
+            stopWhen: stepCountIs(config.maxSteps),
+            providerOptions: this.modelConfig.providerOptions,
+            messages: [
+                {
+                    role: 'system',
+                    content: this.getSystemPrompt(),
+                },
+                ...messages
+            ],
+            ...additionalConfig,
+        })
     }
 
     private getSystemPromptFromType(chatType: ChatType): string {
