+++ packages/ai/src/agents/index.ts
@@ -1,3 +1,2 @@
-export * from './models';
 export * from './classes';
 export { AgentStreamer } from './streamer';
