--- packages/ai/src/agents/models/index.ts
@@ -1,1 +1,0 @@
-export { BaseAgent } from './base';
